Add a generic Tool model Move shared model properties, views and actions from the line and point tools into a common Tool model.
