+++ packages/lib-classifier/src/plugins/drawingTools/models/marks/index.js
@@ -1,2 +1,3 @@
 export { default as Line } from './Line'
+export { default as Mark } from './Mark'
 export { default as Point } from './Point'
