+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/PointTool.js
@@ -1,24 +1,12 @@
 import { types } from 'mobx-state-tree'
+import Tool from './Tool'
 import { Point } from '../marks'
 
 const PointTool = types.model('Point', {
-  color: types.optional(types.string, ''),
-  label: types.optional(types.string, ''),
   marks: types.map(Point),
-  max: types.optional(types.union(types.string, types.number), Infinity),
-  min: types.optional(types.union(types.string, types.number), 0),
   size: types.optional(types.enumeration(['large', 'small']), 'large'),
   type: types.literal('point')
 })
-  .views(self => ({
-    get disabled () {
-      return self.marks.size >= self.max
-    },
-
-    get isComplete () {
-      return (self.marks.size >= self.min)
-    }
-  }))
   .actions(self => {
     function createMark (mark) {
       const newMark = Point.create(mark)
@@ -26,14 +14,9 @@ const PointTool = types.model('Point', {
       return newMark
     }
 
-    function deleteMark (mark) {
-      self.marks.delete(mark.id)
-    }
-
     return {
-      createMark,
-      deleteMark
+      createMark
     }
   })
 
-export default PointTool
+export default types.compose('PointTool', Tool, PointTool)
